+++ src/widgets/toolbarAction.tsx
@@ -14,6 +14,7 @@ import { AuthoringState } from '../editor/authoringState';
 import { DropdownMenuItem, useInsideDropdown } from './utility/dropdown';
 
 import { useWorkspace } from '../workspace/workspaceContext';
+import { EventObserver } from '../workspace';
 
 const CLASS_NAME = 'reactodia-toolbar-action';
 
@@ -342,11 +343,14 @@ export interface ToolbarActionUndoProps extends Omit<ToolbarActionStyleProps, 'd
 /**
  * Toolbar action component to undo a command from the command history.
  *
+ * When mounted, handles the following keyboard shortcuts:
+ *  -  `Ctrl+Z` / `⌘+Z`: perform the undo action.
+ *
  * @category Components
  */
 export function ToolbarActionUndo(props: ToolbarActionUndoProps) {
     const {className, title, ...otherProps} = props;
-    const {model: {history}} = useCanvas();
+    const {canvas, model: {history}} = useCanvas();
     const t = useTranslation();
     const insideDropdown = useInsideDropdown();
     const undoCommand = useObservedProperty(
@@ -358,15 +362,33 @@ export function ToolbarActionUndo(props: ToolbarActionUndoProps) {
                 ? undefined : undoStack[undoStack.length - 1];
         }
     );
+    React.useEffect(() => {
+        const listener = new EventObserver();
+        listener.listen(canvas.events, 'keydown', e => {
+            if (
+                e.sourceEvent.key === 'z' &&
+                (e.sourceEvent.ctrlKey || e.sourceEvent.metaKey) &&
+                !e.sourceEvent.altKey
+            ) {
+                e.sourceEvent.preventDefault();
+                history.undo();
+            }
+        });
+        return () => listener.stopListening();
+    }, [history]);
     const commandTitle = !title && undoCommand ? resolveCommandTitle(undoCommand, t) : undefined;
+    const shortcut = ' (Ctrl+Z / ⌘+Z)';
     return (
         <ToolbarAction {...otherProps}
             className={cx(className, `${CLASS_NAME}__undo`)}
             disabled={!undoCommand}
             title={title ?? (
                 commandTitle === undefined
-                    ? t.text('toolbar_action.undo.title')
-                    : t.text('toolbar_action.undo.title_named', {command: commandTitle})
+                    ? t.text('toolbar_action.undo.title') + shortcut
+                    : t.text(
+                        'toolbar_action.undo.title_named',
+                        {command: commandTitle}
+                    ) + shortcut
             )}
             onSelect={() => history.undo()}>
             {insideDropdown ? t.text('toolbar_action.undo.label') : null}
@@ -384,11 +406,14 @@ export interface ToolbarActionRedoProps extends Omit<ToolbarActionStyleProps, 'd
 /**
  * Toolbar action component to redo a command from the command history.
  *
+ * When mounted, handles the following keyboard shortcuts:
+ *  -  `Ctrl+Shift+Z` / `⌘+Shift+Z`: perform the redo action.
+ *
  * @category Components
  */
 export function ToolbarActionRedo(props: ToolbarActionRedoProps) {
     const {className, title, ...otherProps} = props;
-    const {model: {history}} = useCanvas();
+    const {canvas, model: {history}} = useCanvas();
     const t = useTranslation();
     const insideDropdown = useInsideDropdown();
     const redoCommand = useObservedProperty(
@@ -400,15 +425,33 @@ export function ToolbarActionRedo(props: ToolbarActionRedoProps) {
                 ? undefined : redoStack[redoStack.length - 1];
         }
     );
+    React.useEffect(() => {
+        const listener = new EventObserver();
+        listener.listen(canvas.events, 'keydown', e => {
+            if (
+                e.sourceEvent.key === 'Z' &&
+                (e.sourceEvent.ctrlKey || e.sourceEvent.metaKey) &&
+                !e.sourceEvent.altKey
+            ) {
+                e.sourceEvent.preventDefault();
+                history.redo();
+            }
+        });
+        return () => listener.stopListening();
+    }, [history]);
     const commandTitle = !title && redoCommand ? resolveCommandTitle(redoCommand, t) : undefined;
+    const shortcut = ' (Ctrl+Shift+Z / ⌘+Shift+Z)';
     return (
         <ToolbarAction {...otherProps}
             className={cx(className, `${CLASS_NAME}__redo`)}
             disabled={!redoCommand}
             title={title ?? (
                 commandTitle === undefined
-                    ? t.text('toolbar_action.redo.title')
-                    : t.text('toolbar_action.redo.title_named', {command: commandTitle}) 
+                    ? t.text('toolbar_action.redo.title') + shortcut
+                    : t.text(
+                        'toolbar_action.redo.title_named',
+                        {command: commandTitle}
+                    ) + shortcut 
             )}
             onSelect={() => history.redo()}>
             {insideDropdown ? t.text('toolbar_action.redo.label') : null}